Выполняет шаги из задания - Учебный проект: время разбивать камни
