+++ public/index.html
@@ -1923,6 +1923,7 @@ Here is a card with filled data</textarea
       </section>
     </main>
 
-    <script type="text/javascript" src="../src/main.js"></script>
+    <!-- <script type="text/javascript" src="../src/main.js"></script> -->
+    <script type="text/javascript" src="bundle.js"></script>
   </body>
 </html>
