+++ src/main.js
@@ -1,191 +1,8 @@
-'use strict';
+import {renderSeveralCard, START_COUNT_CARD} from './card';
+import {FilterType, renderFilter} from './filter';
 
-(function () {
-  const MAX_NUMBER_TASKS = 100;
+for (let i = 0, l = FilterType.length; i < l; i++) {
+  renderFilter(FilterType[i]);
+}
 
-  const START_COUNT_CARD = 7;
-  const MAX_COUNT_CARD = 20;
-
-  const mainFiltersBlock = document.querySelector(`.main__filter`);
-  mainFiltersBlock.innerHTML = ``;
-
-  const FilterType = [
-    `ALL`,
-    `OVERDUE`,
-    `TODAY`,
-    `FAVORITES`,
-    `REPEATING`,
-    `TAGS`,
-    `ARCHIVE`
-  ];
-
-  const FilterTemplate = {
-    ALL: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__all"
-        class="filter__input visually-hidden"
-        name="filter"
-        checked
-      />
-      <label for="filter__all" class="filter__label">
-        ALL <span class="filter__all-count">${randomNumberTask}</span></label
-      >
-    `,
-    OVERDUE: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__overdue"
-        class="filter__input visually-hidden"
-        name="filter"
-        disabled
-      />
-      <label for="filter__overdue" class="filter__label"
-        >OVERDUE <span class="filter__overdue-count">${randomNumberTask}</span></label
-      >
-    `,
-    TODAY: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__today"
-        class="filter__input visually-hidden"
-        name="filter"
-        disabled
-      />
-      <label for="filter__today" class="filter__label"
-        >TODAY <span class="filter__today-count">${randomNumberTask}</span></label
-      >
-    `,
-    FAVORITES: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__favorites"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__favorites" class="filter__label"
-        >FAVORITES <span class="filter__favorites-count">${randomNumberTask}</span></label
-      >
-    `,
-    REPEATING: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__repeating"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__repeating" class="filter__label"
-        >Repeating <span class="filter__repeating-count">${randomNumberTask}</span></label
-      >
-    `,
-    TAGS: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__tags"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__tags" class="filter__label"
-        >Tags <span class="filter__tags-count">${randomNumberTask}</span></label
-      >
-    `,
-    ARCHIVE: (randomNumberTask) => `
-      <input
-        type="radio"
-        id="filter__archive"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__archive" class="filter__label"
-        >ARCHIVE <span class="filter__archive-count">${randomNumberTask}</span></label
-      >
-    `
-  };
-
-  const getFilterTemplate = (filterType) => {
-    const randomNumberTask = Math.round(Math.random() * MAX_NUMBER_TASKS);
-
-    return FilterTemplate[filterType](randomNumberTask);
-  };
-
-  const renderFilter = (filterType) => {
-    const wrapperFilter = document.createElement(`div`);
-    wrapperFilter.innerHTML = getFilterTemplate(filterType);
-
-    wrapperFilter.addEventListener(`click`, () => {
-      renderSeveralCard();
-    });
-
-    mainFiltersBlock.appendChild(wrapperFilter);
-  };
-
-  const boardTasksBlock = document.querySelector(`.board__tasks`);
-  boardTasksBlock.innerHTML = ``;
-
-  const renderCard = () => {
-    const card = document.createElement(`article`);
-    card.className = `card`;
-    card.innerHTML = `
-      <form class="card__form" method="get">
-        <div class="card__inner">
-          <div class="card__control">
-            <button type="button" class="card__btn card__btn--edit">
-              edit
-            </button>
-            <button type="button" class="card__btn card__btn--archive">
-              archive
-            </button>
-            <button type="button" class="card__btn card__btn--favorites card__btn--disabled">
-              favorites
-            </button>
-          </div>
-          <div class="card__color-bar">
-            <svg class="card__color-bar-wave" width="100%" height="10">
-              <use xlink:href="#wave"></use>
-            </svg>
-          </div>
-          <div class="card__textarea-wrap">
-            <label>
-              <textarea class="card__text" placeholder="Start typing your text here..." name="text"></textarea>
-            </label>
-          </div>
-          <div class="card__settings">
-            <div class="card__details">
-              <div class="card__dates"></div>
-              <div class="card__hashtag">
-                <div class="card__hashtag-list"></div>
-                <label>
-                  <input type="text" class="card__hashtag-input" name="hashtag-input" placeholder="Type new hashtag here">
-                </label>
-              </div>
-            </div>
-            <label class="card__img-wrap card__img-wrap--empty">
-              <input type="file" class="card__img-input visually-hidden" name="img">
-              <img src="img/add-photo.svg" alt="task picture" class="card__img">
-            </label>
-            <div class="card__colors-inner">
-              <h3 class="card__colors-title">Color</h3>
-              <div class="card__colors-wrap"></div>
-            </div>
-          </div>
-        </div>
-      </form>`;
-
-    boardTasksBlock.appendChild(card);
-  };
-
-  const renderSeveralCard = (number = Math.round(Math.random() * MAX_COUNT_CARD)) => {
-    boardTasksBlock.innerHTML = ``;
-
-    for (let i = 0; i < number; i++) {
-      renderCard();
-    }
-  };
-
-  for (let i = 0, l = FilterType.length; i < l; i++) {
-    renderFilter(FilterType[i]);
-  }
-
-  renderSeveralCard(START_COUNT_CARD);
-
-}());
+renderSeveralCard(START_COUNT_CARD);
